requirements.txt
python-social-auth
requests_oauthlib
gunicorn
django
requests
